SpeedyWeatherEmulator

Documentation for SpeedyWeatherEmulator.

SpeedyWeatherEmulator.BasicStructs.SimPara

SpeedyWeatherEmulator.BasicStructs.SimPara

SpeedyWeatherEmulator.SimDataHandling.SimData

SpeedyWeatherEmulator.SimDataHandling.SimData

SpeedyWeatherEmulator.SimDataHandling.create_sim_DIR

SpeedyWeatherEmulator.SimDataHandling.load_sim_data

SpeedyWeatherEmulator.SimDataHandling.save_sim_data

SpeedyWeatherEmulator.BasicStructs.SimPara — Type

SimPara(trunc::Int, 
        n_steps::Int, 
        n_ic::Int, 
        n_spinup::Int, 
        t_step::Real, 
        initial_cond::Union{Nothing, Function}, 
        id_key::String)

Container for basic simulation parameters for SpeedyWeather.jl simulation data generation and data shape, see constructor for details.

SpeedyWeatherEmulator.BasicStructs.SimPara — Method

SimPara(;   trunc::Int, 
            n_steps::Int, 
            n_ic::Int, 
            n_spinup::Int = 10, 
            t_step::Real = 1.0, 
            initial_cond::Union{Nothing, Function} = nothing, 
            id_key::String = "")

Convenience Constructor for SimPara

Fields

trunc::Int: Spectral truncation of the barotropic model (e.g. 5 for T5).

n_steps::Int: Number of data points after the spin-up time, with the first at t = t_spinup.

n_ic::Int: Number of simulated initial conditions.

n_spinup::Int = 10: Time after data is stored in SimData

t_step::Real = 1.0: Time step length.

initial_cond::Union{Nothing, Function} = nothing: Specific initial condition, if empty: random initial conditions are applied.

id_key::String = "": Key for accessing saved simulation data. Data is saved with the parameters trunc, nsteps and nic. Key is for further differentiation of simulation data.

Examples

```julia simpara = SimPara(trunc=5, nsteps=8, nic=1000, storagekey="test")

SpeedyWeatherEmulator.SimDataHandling.SimData — Type

SimData(sim_para::SimPara, 
        data::Array{Float32,3})

Container for SpeedyWeather.jl simulation data and corresponding simulation parameter, see constructor for details.

Fields

sim_para::SimPara: Simulation parameters for data generation.

data::Array{Float32,3}: Data storage, where data = (ncoeff) x `nstepsxnic`, where (ncoeff) is the number of spectral coefficients (real and imaginary part), e.g. for trunc=5: n_coeff=54

SpeedyWeatherEmulator.SimDataHandling.SimData — Method

SimData(sim_para::SimPara)

Convenience Constructor for SimData and vorticity data generation.

Corresponding to the simulation parameters in SimPara, this constructor: - defines a SpeedyWeather.jl simulation sim - runs it for different random initial conditions (if there is no other initial condition defined in SimPara) - stores the vorticity of the actual leap frog step in data, for array details see struct definition

Arguments

sim_para::SimPara: Simulation parameters for data generation.

Examples

sim_para = SimPara(trunc=5, n_steps=8, n_ic=1000)
sim_data = SimData(sim_para)

SpeedyWeatherEmulator.SimDataHandling.create_sim_DIR — Method

create_sim_DIR(;sim_para::SimPara)

Creates a DIR for saving and loading SimulationData according to sim_para.

Arguments

sim_para::SimPara: Simulation parameters (key) for the DIR generation.

Returns

filepath::String: Returns the filepath

SpeedyWeatherEmulator.SimDataHandling.load_sim_data — Method

load_sim_data(sim_para::SimPara)

Loads existing simulation data from data/simdata/ according to simdata.sim_para.

Arguments

sim_para::SimPara: Simulation parameters (key) of the stored simulation data

Returns

SimData: Returns the loaded simulation data

Examples

sim_para = SimPara(trunc=5, n_steps=8, n_ic=1000)
sim_data = SimData(sim_para)
save_sim_data(sim_data)
...
sim_data_loaded = load_sim_data(sim_para)

SpeedyWeatherEmulator.SimDataHandling.save_sim_data — Method

save_sim_data(sim_data::SimData)

Saves simulation data in data/simdata/ according to simdata.sim_para.

Arguments

sim_data::SimData: Storaged data and simulation parameters

Returns

nothing

Examples

sim_para = SimPara(trunc=5, n_steps=8, n_ic=1000)
sim_data = SimData(sim_para)
save_sim_data(sim_data)